Filters
Added filters for all fields except userPerforming because I don't know if that should be a sortable field. Also commented out the get all transactions method for now because the filters are optional so we can use that one. But I will ask if there is a reason to keep a separate select all option